+++ MyPaint/MyPaint/CustomControl/PopupTextBox.cs
