+++ MyPaint/MyPaint/PaintFormer/ShapeFormer.cs
